feat: useControlStore and ControlStatus separated into two files
