+++ src/components/react-components/model-editor/StartCompTabContent/ControlCompParams/ControlCompParams.tsx
@@ -1,4 +1,4 @@
-import useControlStore from '../../../../../stores/LiveModel/useControlStore';
+import useControlStore from '../../../../../stores/LiveModel/ControlStore/useControlStore';
 import type { ControlInfo } from '../../../../../types';
 import DotHeaderReact from '../../../lit-wrappers/DotHeaderReact';
 import NumberInputReact from '../../../lit-wrappers/NumberInputReact';
