+++ src/services/global/LiveModel/ExportLM.ts
@@ -1,3 +1,11 @@
+import { Message } from '../../../components/lit-components/message-wrapper';
+import config from '../../../config';
+import useControlStore from '../../../stores/LiveModel/ControlStore/useControlStore';
+import useLoadedModelStore from '../../../stores/LiveModel/useLoadedModelStore';
+import useModelInfoStore from '../../../stores/LiveModel/useModelInfoStore';
+import useRegulationsStore from '../../../stores/LiveModel/useRegulationsStore';
+import useUpdateFunctionsStore from '../../../stores/LiveModel/useUpdateFunctionsStore';
+import useVariablesStore from '../../../stores/LiveModel/useVariablesStore';
 import type {
   ControlInfo,
   fileType,
@@ -5,16 +13,8 @@ import type {
   Variable,
 } from '../../../types';
 import CytoscapeME from '../../model-editor/ModelVisualization/CytoscapeME';
-import type { LiveModelClass } from './LiveModel';
-import useVariablesStore from '../../../stores/LiveModel/useVariablesStore';
-import useUpdateFunctionsStore from '../../../stores/LiveModel/useUpdateFunctionsStore';
-import useRegulationsStore from '../../../stores/LiveModel/useRegulationsStore';
-import useControlStore from '../../../stores/LiveModel/useControlStore';
-import useModelInfoStore from '../../../stores/LiveModel/useModelInfoStore';
-import config from '../../../config';
-import { Message } from '../../../components/lit-components/message-wrapper';
-import useLoadedModelStore from '../../../stores/LiveModel/useLoadedModelStore';
 import FileHelpers from '../../utilities/FileHelpers';
+import type { LiveModelClass } from './LiveModel';
 
 class ExportLM {
   // #region --- Properties + Constructor ---
